+++ HISTORY.md
@@ -4,7 +4,9 @@
 
 ## [unreleased/patches]
 
+- feat: player gets message if area is renamed [0.1.4]
 - refactor: now `/lc` opens the menu too (instead of doing nothing) [0.1.2]
+- refactor: default area permissions adjusted [0.1.4]
 - fix: missing translation for changing player permissions of a zone [0.1.1]
 - fix: expanding area did not take origin permisssions into account [0.1.3]
 - fix: Dropdown z-index issue workaround in PermissionPanel [0.1.3]